more foundations for the new tracker and changed user id to i32 (rust) and INT (sql)

+++ backend/api/src/middlewares/jwt_middleware.rs
@@ -15,7 +15,7 @@ use jsonwebtoken::{decode, errors::ErrorKind, DecodingKey, Validation};
 
 #[derive(Debug, Clone)]
 pub struct Authdata {
-    pub sub: i64,
+    pub sub: i32,
     pub class: UserClass,
 }
