+++ backend/storage/src/models/torrent_report.rs
@@ -8,7 +8,7 @@ pub struct TorrentReport {
     pub id: i64,
     #[schema(value_type = String, format = DateTime)]
     pub reported_at: DateTime<Local>,
-    pub reported_by_id: i64,
+    pub reported_by_id: i32,
     pub reported_torrent_id: i64,
     pub description: String,
 }
