+++ backend/storage/src/models/torrent_request.rs
@@ -25,8 +25,8 @@ pub struct TorrentRequest {
     pub created_at: DateTime<Utc>,
     #[schema(value_type = String, format = DateTime)]
     pub updated_at: DateTime<Utc>,
-    pub created_by_id: i64,
-    pub filled_by_user_id: Option<i64>,
+    pub created_by_id: i32,
+    pub filled_by_user_id: Option<i32>,
     pub filled_by_torrent_id: Option<i64>,
     #[schema(value_type = String, format = DateTime)]
     pub filled_at: Option<DateTime<Utc>>,
