+++ backend/storage/src/repositories/edition_group_repository.rs
@@ -9,10 +9,10 @@ impl ConnectionPool {
     pub async fn create_edition_group(
         &self,
         edition_group_form: &UserCreatedEditionGroup,
-        current_user_id: i64,
+        current_user_id: i32,
     ) -> Result<EditionGroup> {
         const CREATE_EDITION_GROUPS_QUERY: &str = r#"
-            INSERT INTO edition_groups (title_group_id, name, release_date, created_by_id, description, distributor, covers, external_links, source, additional_information) 
+            INSERT INTO edition_groups (title_group_id, name, release_date, created_by_id, description, distributor, covers, external_links, source, additional_information)
             VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9::source_enum, $10)
             RETURNING *;
         "#;
